+++ src/Data/Extensible/Label.hs
@@ -27,11 +27,6 @@ instance k ~ l => IsLabel k (Proxy l) where
   fromLabel _ = Proxy
 #endif
 
--- | Specialised version of 'itemAssoc'.
-訊 :: Proxy k -> FieldOptic k
-訊 = itemAssoc
-{-# DEPRECATED 訊 "Use xlb instead" #-}
-
 -- | Specialised version of 'itemAssoc'. Stands for "eXtensible LaBel"
 xlb :: Proxy k -> FieldOptic k
 xlb = itemAssoc